ci: shard long-running test jobs more aggressively
Workspace tests, runner tests, and Pattern Integration tests have become
the long poles in CI. Increase their matrix sizes so their work can run
in parallel across more jobs.

Split the large CLI commit-message suite into independent physical test
files. Keep them in the serial inventory because they change shared Git
configuration and process state, while allowing ordinary sorted shard
assignment to distribute them across the CLI slices. Refresh the CLI
timing weights for the resulting layout.

Assign measured Pattern Integration files with a profiled table that
keeps expensive files apart and leaves the first shard available for
heavier internally sharded work. Use sorted round-robin assignment for
new files that have not been profiled. Include the shard selector in the
compile-cache key so each job compiles the sources it will execute.

Keep coverage artifact checks and CI contract tests aligned with the
expanded matrices. Make active performance guidance refer to configured
matrices and numbered test families without repeating shard counts.
