+++ packages/patterns/deno.jsonc
@@ -28,8 +28,8 @@
   // Lane 3 — integration test. A `*.test.ts` file under `integration/` that
   // needs a running shell/toolshed/browser (PiecesController, ShellIntegration,
   // end-to-end flows). It is run by `deno task integration`. CI runs these in
-  // the "Pattern Integration Tests" job across four shards; a heavy new file
-  // can be pinned to a shard in tasks/select-pattern-integration-files.ts, and
+  // the "Pattern Integration Tests" job across its shard matrix; a heavy new
+  // file can be assigned in tasks/select-pattern-integration-files.ts, and
   // otherwise falls back to round-robin. Reload-specific cases live in
   // integration/reload/ and run via `integration:reload`.
   //
