+++ tasks/ci-workflow.test.ts
@@ -1,4 +1,6 @@
 import { assert, assertEquals, assertStringIncludes } from "@std/assert";
+import { EXPECTED_COVERAGE_ARTIFACT_NAMES } from "./coverage-check.ts";
+import { PATTERN_INTEGRATION_SHARD_COUNT } from "./select-pattern-integration-files.ts";
 
 function jobBlock(workflow: string, jobId: string): string {
   const jobsStart = workflow.indexOf("jobs:\n");
@@ -132,6 +134,63 @@ Deno.test("Coverage Comment follows the CI workflow by name", async () => {
   assert(name, "workflow name not found");
 
   assertStringIncludes(comment, `    workflows: ["${name[1]}"]\n`);
+});
+
+Deno.test("coverage requirements follow sharded test matrices", async () => {
+  const contents = await workflow("deno.yml");
+  const jobs = [
+    ["test", "coverage-profile-workspace-"],
+    ["runner-test", "coverage-profile-runner-"],
+    ["pattern-integration-test", "coverage-profile-pattern-integration-"],
+  ] as const;
+
+  for (const [jobId, artifactPrefix] of jobs) {
+    const job = jobBlock(contents, jobId);
+    const shardMatch = job.match(/^ {8}shard: \[([0-9, ]+)\]$/m);
+    const totalMatch = job.match(/^ {8}total: \[(\d+)\]$/m);
+    assert(shardMatch, `${jobId} shard matrix not found`);
+    assert(totalMatch, `${jobId} total matrix not found`);
+
+    const shards = shardMatch[1].split(",").map((value) =>
+      Number(value.trim())
+    );
+    const total = Number(totalMatch[1]);
+    if (jobId === "pattern-integration-test") {
+      assertEquals(
+        total,
+        PATTERN_INTEGRATION_SHARD_COUNT,
+        "pattern integration matrix must use its measured assignment count",
+      );
+    }
+    assertEquals(
+      shards,
+      Array.from({ length: total }, (_, index) => index + 1),
+      `${jobId} must list every shard exactly once`,
+    );
+    assertStringIncludes(
+      job,
+      `name: ${artifactPrefix}\${{ matrix.shard }}`,
+    );
+    assertEquals(
+      EXPECTED_COVERAGE_ARTIFACT_NAMES.filter((name) =>
+        name.startsWith(artifactPrefix)
+      ),
+      shards.map((shard) => `${artifactPrefix}${shard}`),
+      `${jobId} coverage requirements must match its matrix`,
+    );
+  }
+});
+
+Deno.test("pattern integration cache follows its shard topology", async () => {
+  const contents = await workflow("deno.yml");
+  const job = jobBlock(contents, "pattern-integration-test");
+  const topology = "-${{ matrix.total }}-${{ matrix.shard }}-";
+
+  assertEquals(job.split(topology).length - 1, 2);
+  assertStringIncludes(job, "'tasks/select-pattern-integration-files.ts'");
+  const restoreKeys = job.match(/restore-keys: \|\n((?: {12}.+\n)+)/);
+  assert(restoreKeys, "pattern integration restore prefixes not found");
+  assertEquals(restoreKeys[1].trim().split("\n").length, 1);
 });
 
 Deno.test("Dashboard publishes only from main, never from a pull request", async () => {
