+++ tasks/coverage-check.ts
@@ -1060,16 +1060,20 @@ export async function collectCurrentCacheStates(
   }
 }
 
-const EXPECTED_COVERAGE_ARTIFACT_NAMES = [
-  ...[1, 2, 3, 4, 5, 6].map((shard) => `coverage-profile-workspace-${shard}`),
-  ...[1, 2, 3, 4, 5].map((shard) => `coverage-profile-runner-${shard}`),
+export const EXPECTED_COVERAGE_ARTIFACT_NAMES = [
+  ...[1, 2, 3, 4, 5, 6, 7, 8, 9, 10].map((shard) =>
+    `coverage-profile-workspace-${shard}`
+  ),
+  ...[1, 2, 3, 4, 5, 6, 7, 8, 9, 10].map((shard) =>
+    `coverage-profile-runner-${shard}`
+  ),
   ...[1, 2, 3, 4].map((shard) =>
     `coverage-profile-generated-patterns-${shard}`
   ),
   "coverage-profile-package-runner",
   "coverage-profile-package-runtime-client",
   "coverage-profile-package-shell",
-  ...[1, 2, 3, 4].map((shard) =>
+  ...[1, 2, 3, 4, 5, 6, 7, 8].map((shard) =>
     `coverage-profile-pattern-integration-${shard}`
   ),
   "coverage-profile-pattern-reload",
