+++ tasks/test-timing-weights.ts
@@ -97,9 +97,16 @@ export const WORKSPACE_TEST_WEIGHTS: Readonly<Record<string, number>> = {
   api: 2.3,
   "background-piece-service": 7.2,
   "cf-harness": 29.9,
-  "cli (1/3)": 92.7,
-  "cli (2/3)": 33.6,
-  "cli (3/3)": 70.5,
+  "cli (1/10)": 50.5,
+  "cli (2/10)": 71,
+  "cli (3/10)": 37,
+  "cli (4/10)": 31,
+  "cli (5/10)": 18.5,
+  "cli (6/10)": 24,
+  "cli (7/10)": 46,
+  "cli (8/10)": 39,
+  "cli (9/10)": 25,
+  "cli (10/10)": 24.5,
   "content-hash": 3.6,
   dashboard: 14.8,
   "data-model": 9.3,
