+++ tasks/workspace-tests.ts
@@ -138,7 +138,7 @@ const INTERNALLY_SHARDED_PACKAGES: Record<
   { total: number; envVar: string }
 > = {
   // packages/cli/test/run-tests.ts reads CLI_TEST_SHARD.
-  cli: { total: 3, envVar: "CLI_TEST_SHARD" },
+  cli: { total: 10, envVar: "CLI_TEST_SHARD" },
   piece: { total: 3, envVar: "PIECE_TEST_SHARD" },
   tasks: { total: 3, envVar: "TASK_TEST_SHARD" },
 };
